Created types directory and renamed types.ts - Renamed types.ts to user-types.ts to hold interfaces relating to users - Refactor components to utilize new directory and file for types.ts
