+++ src/components/EditProfile.tsx
@@ -4,7 +4,7 @@ import {
 } from 'react-native';
 
 import Input from './Input';
-import type { IUser } from '../types';
+import type { IUser } from '../types/users-types';
 
 interface EditProfileProp {
     loggedUser: IUser;              // User currently signed in the app
