+++ src/screens/UserProfile.tsx
@@ -9,7 +9,7 @@ import {
 } from 'react-native-paper';
 
 import InfoBox from '../components/InfoBox';
-import type { IUser } from '../types';
+import type { IUser } from '../types/users-types';
 
 interface UserProfileProp {
     user: IUser;
